requirements.txt
annotated-types==0.7.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fonttools==4.59.2
hdbscan==0.8.40
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
murmurhash==1.0.13
narwhals==2.5.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.3
